requirements.txt
aiofiles==24.1.0
aiogram==3.18.0
aiogram_dialog==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
amqp==5.3.1
annotated-types==0.7.0
asyncio==3.4.3
attrs==25.3.0
billiard==4.2.1
cachetools==5.5.2
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dataclasses==0.6
datadispatch==1.0.0
Deprecated==1.2.18
distro==1.9.0
environs==14.1.1
frozenlist==1.5.0
idna==3.10
Jinja2==3.1.6
kombu==5.5.0
magic-filter==1.0.12
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.2.0
netaddr==1.3.0
packaging==24.2
prompt_toolkit==3.0.50
propcache==0.3.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
six==1.17.0
SQLAlchemy==2.0.39
tasks==2.8.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
wrapt==1.17.2
yarl==1.18.3
yookassa==3.5.0
